[TRTLLM-13212][refactor] Align sampling backend layout with attention_backend
Address review feedback (naming consistency + a missed call site):

- Rename sampling_backends/ -> sampling_backend/ (singular, matching
  attention_backend/) and drop the redundant backend_ filename prefix:
  backend_torch.py -> vanilla.py, backend_flashinfer.py -> flashinfer.py,
  backend_custom.py -> trtllm.py (the latter wraps torch.ops.trtllm custom
  ops, so trtllm.py is more explicit than "custom").
- Move SamplerConfig / BoundSamplingBackend / resolve_sampling_backend into
  sampling_backend/interface.py (mirrors attention_backend/interface.py).
- Update all imports, attribute accesses, and docstrings accordingly.
- Fix eagle3_dynamic_tree.py: drop the stale use_flashinfer kwarg from the
  context-subset sampling_batch_spec_dec_one_model call (the parameter was
  removed earlier in this refactor); the call now matches the new signature.
